Add the identifiability diagnostic; pin the reference problem's dtype
Several parameters in the gallery cannot be recovered from the observations at
all (Fisher analysis puts sindy c2/c3 at info gain 1.04x/1.02x and SEIRD alpha at
1.06x), yet they were averaged into headline numbers -- which is how "SINDy 18.0%
vs 32.4%, 1.8x better" ended up resting partly on parameters that carry no signal.

controls.contraction() reports prior sd / posterior sd per parameter, and the
scoreboard now prints a verdict per parameter. A flat posterior is ambiguous on
its own, so the ridge control arbitrates:

    NO INFO    posterior is flat AND the ridge cannot beat the prior
               -> the data does not contain the parameter; scoring it is noise
    WE FAILED  posterior is flat but the ridge DOES extract it
               -> the information is there and our model missed it

On sindy_sde this immediately splits the five parameters: c0/c2/c3 are NO INFO
(agreeing with the independent Fisher computation), while c1 and sigma are OUR
failures -- a degree-2 ridge recovers sigma to 7.5% where we return the prior at
22.9%.

Also: linear_gaussian's design matrix used torch.randn without a dtype, so the
"fixed, seeded" reference silently changed with the ambient default dtype (the
posterior became 2.3x wider under float64). The instrument we measure everything
against must not depend on global state; dtype is now pinned.

+++ amortix/problems/linear_gaussian.py
@@ -28,9 +28,13 @@ N_OBS = 6
 NOISE = 0.5
 
 # fixed, correlated design matrix (seeded once, not learned)
+# dtype is pinned: torch.randn follows the ambient default dtype, so this
+# "fixed, seeded" matrix silently changed (and the posterior got 2.3x wider)
+# whenever the process default was float64. The reference instrument must not
+# depend on global state.
 _g = torch.Generator().manual_seed(20240630)
-A = (torch.randn(N_OBS, D_PARAM, generator=_g) * 0.6
-     + torch.linspace(0.4, 1.0, D_PARAM)[None, :])
+A = (torch.randn(N_OBS, D_PARAM, generator=_g, dtype=torch.float32) * 0.6
+     + torch.linspace(0.4, 1.0, D_PARAM, dtype=torch.float32)[None, :])
 
 
 class _VectorObserver:
